diploma-level/modern-application-development-2/BlogApp/requirements.txt
alembic==1.10.2
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
billiard==3.6.4.0
blinker==1.6.2
cachelib==0.9.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Flask==2.2.3
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
pdfkit==1.0.0
prompt-toolkit==3.0.38
PyJWT==2.6.0
pytz==2023.2
redis==4.5.4
six==1.16.0
SQLAlchemy==2.0.7
typing-extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
